Fixed some types used in the c code generator

+++ Datalog/src/C_Template/data_structure.c
@@ -43,7 +43,7 @@ void Ds_get_intValues_Level0_init(){
         first_value = 1;
 }
 
-int Ds_get_intValues_Level0(int *value){
+short Ds_get_intValues_Level0(unsigned int *value){
         Word_t * PValue;
 
         if (first_value){
